Allow <includePluginDependencies> to be specified for the exec:exec goal
