+++ src/it/projects/exec-plugin-dependencies/invoker.properties
@@ -1,0 +1,1 @@
+invoker.goals = exec:exec
